refactor(kv): inline the single-runtime KV factory into kv-redis
The Cloudflare KV backend is gone, so runtime/factory.ts only forwarded to
the Redis constructor. Export createKVStoreSet from kv-redis directly and drop
the passthrough module.
