+++ apps/sdp-api/src/runtime/kv-redis.node.test.ts
@@ -8,7 +8,7 @@
 import Redis from "ioredis";
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it } from "vitest";
 import type { Env } from "@/types/env";
-import { closeAllRedisClients, createRedisKVStoreSet, RedisKVStore } from "./kv-redis";
+import { closeAllRedisClients, createKVStoreSet, RedisKVStore } from "./kv-redis";
 
 const REDIS_URL = process.env.REDIS_URL ?? "redis://127.0.0.1:6379";
 
@@ -107,14 +107,14 @@ describe("RedisKVStore (HOO-510)", () => {
     });
   });
 
-  describe("createRedisKVStoreSet", () => {
+  describe("createKVStoreSet", () => {
     afterEach(async () => {
       // Factory uses a cached client; close between tests so reuse is observable.
       await closeAllRedisClients();
     });
 
     it("wires four prefixed stores sharing one connection", async () => {
-      const set = createRedisKVStoreSet({ REDIS_URL } as Env);
+      const set = createKVStoreSet({ REDIS_URL } as Env);
       await set.apiKeys.put("same-key", "from-apiKeys");
       await set.rateLimits.put("same-key", "from-rateLimits");
       await set.cache.put("same-key", "from-cache");
@@ -129,21 +129,19 @@ describe("RedisKVStore (HOO-510)", () => {
     it("reuses the same Redis client across repeated calls (no connection leak)", () => {
       // Pin reference equality on the cached promise. If the factory ever
       // regresses to per-call construction, the connection-leak comes back.
-      const set1 = createRedisKVStoreSet({ REDIS_URL } as Env);
-      const set2 = createRedisKVStoreSet({ REDIS_URL } as Env);
+      const set1 = createKVStoreSet({ REDIS_URL } as Env);
+      const set2 = createKVStoreSet({ REDIS_URL } as Env);
       const p1 = (set1.apiKeys as unknown as { clientPromise: Promise<Redis> }).clientPromise;
       const p2 = (set2.apiKeys as unknown as { clientPromise: Promise<Redis> }).clientPromise;
       expect(p1).toBe(p2);
     });
 
     it("throws a clear error when REDIS_URL is missing", () => {
-      expect(() => createRedisKVStoreSet({} as Env)).toThrow(/REDIS_URL is required/);
+      expect(() => createKVStoreSet({} as Env)).toThrow(/REDIS_URL is required/);
     });
 
     it("throws when REDIS_URL is whitespace-only", () => {
-      expect(() => createRedisKVStoreSet({ REDIS_URL: "   " } as Env)).toThrow(
-        /REDIS_URL is required/
-      );
+      expect(() => createKVStoreSet({ REDIS_URL: "   " } as Env)).toThrow(/REDIS_URL is required/);
     });
   });
 });
