+++ apps/sdp-api/src/test/integration-support.ts
@@ -6,8 +6,7 @@ import { hashString } from "@sdp/payments/hash";
 
 import { closeDatabasePools, getDb } from "@/db";
 import app from "@/index";
-import { createKVStoreSet } from "@/runtime/factory";
-import { closeAllRedisClients } from "@/runtime/kv-redis";
+import { closeAllRedisClients, createKVStoreSet } from "@/runtime/kv-redis";
 import { createFeePaymentAdapter, KoraAdapter, KoraClient } from "@/services/adapters";
 import { createSigningService } from "@/services/domain/signing.service";
 import { createMosaicService } from "@/services/mosaic";
